requirements.txt
ultralytics
streamlit
pillow
requests
torch
torchvision
opencv-python-headless
numpy